backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.7
python-multipart>=0.0.6
requests>=2.28.2
Pillow>=9.5.0
numpy>=1.24.3
aiofiles>=23.1.0
pytest>=7.3.1
python-dotenv>=1.0.0
jsonrpc>=1.2.0
sse-starlette>=1.6.1
